add a method to check if the list belongs to the user

+++ Back/app/controllers/listHasBookController.js
@@ -6,6 +6,16 @@ const listHasbookMapper = require('../models/listHasBookMapper');
 const listHasbookController = {
   addBookToList: async (req, res) => {
     try {
+      req.body.user_id = req.user.id;
+      const ListBelongsToUser = await listHasbookMapper.checkIfListBelongsToUser(
+        req.body
+      );
+      if (!ListBelongsToUser)
+        return res
+          .status(400)
+          .json(
+            `The list ${req.body.list_id} does not belong to the user ${req.body.user_id}`
+          );
       // We check if the book is in the database
       const book = await bookMapper.getBookByPublicApiId(
         req.body.public_api_id
@@ -34,6 +44,15 @@ const listHasbookController = {
   deleteBookFromList: async (req, res) => {
     req.body.user_id = req.user.id;
     try {
+      const ListBelongsToUser = await listHasbookMapper.checkIfListBelongsToUser(
+        req.body
+      );
+      if (!ListBelongsToUser)
+        return res
+          .status(400)
+          .json(
+            `The list ${req.body.list_id} does not belong to the user ${req.body.user_id}`
+          );
       await listHasbookMapper.deleteBookFromList(req.body);
       res.json('OK');
     } catch (error) {
